egraph: Ensure eclass members are all available in the same block (#8006)
* egraph: Ensure eclass members are all available in the same block

During elaboration, we want the property that it's safe to pick any
member of an eclass to implement that eclass. To establish that
property, this commit defines a notion of the "available block" for
every value, and enforces that all values in an eclass have the same
available block.

Closes #7891 where we devised this plan.

Co-authored-by: L Pereira <l.pereira@fastly.com>

* Review comments

- add documentation comments
- push orig_value onto optimized_values before calling simplify so it
  can't cause a heap reallocation

* Delay pushing orig_value so filetests don't change

* Don't set available-block for instructions removed by GVN

Because the duplicate instruction is never used, it never matters what
block it might be available in. For the same reason it also doesn't need
to be in the union-find.

* Add tests

---------

Co-authored-by: L Pereira <l.pereira@fastly.com>