x64: Implement fallback for fcvt_from_sint.f64x2
This instruction was previously unimplemented in Cranelift as pointed
out in #8084. While a fallback for `fcvt_from_uint` was implemented
in #7919 I forgot to do the same for the signed version. This commit
adds a fallback that decomposes the input into two scalars and converts
each individually, then reassembling the result.