Add a script to vendor WIT files
This is a follow-up to #7929 where it was concluded that we want to
retain textual WIT files in this repository. Currently today we have no
automated checks for vendoring and re-vendoring requires hand-editing.
To help incrementally improve this situation this commit adds a script
to do all the vendoring for us. This way CI can verify that everything
is generated with respect to the output of this script.