Stop threading Cranelift tables through cranelift-wasm callbacks (#8105)
* Stop threading Cranelift tables through cranelift-wasm callbacks

Previously, every table-related method in the FuncEnvironment trait took
two IDs to identify a table: the TableIndex from the WebAssembly module,
and the Cranelift Table we generated associated with it.

This is redundant and most implementations only used one or the other.
The TableIndex is a perfectly good identifier and we can use it in a
SecondaryMap if we need to associate additional data with the table,
such as a Cranelift Table ID.

This is a step toward #5532, where we don't want to use Cranelift's
table support at all, so the associated data that an implementation
needs will be different.

* Fix unused variable warning in non-GC configurations