x64: Fix load sinking in `rounds{s,d}` loading too much
This commit aims to address #8116 by fixing these two instructions to
load the proper amount of bytes when a load is sunk into them. Currently
the instruction variant used here requires an aligned `XmmMem` which
is today auto-translated with a 16-byte load unconditionally. For loads
near the end of memory this loads too much and can erroneously cause a
trap. The fix in this commit is to force the load to happen manually
with the appropriate type rather than a 16-byte type.

I'll note that `XmmMem` is left as an argument in this case because the
AVX variants of these instructions can continue to leverage unaligned
accesses. This also means that the test added here won't fail on a
machine with AVX support, it needs to be explicitly disabled.

Closes #8116