cranelift-wasm: Emit constant bounds for fixed-size tables (#8125)
WebAssembly tables have a minimum size, but may optionally also have a
maximum size. If the maximum is set to the same number of elements as
the minimum, then we can emit an `iconst` instruction for bounds-checks
on tables, rather than loading the bounds from memory.

That's a win in its own right, but when optimization is enabled, this
also allows bounds-checks to constant-fold if the table index is
provided as an integer constant.