s390x: Fix ABI vector-lane swap for overflow args
On s390x we support ABIs with little-endian or big-endian vector
lane ordering.  When calling to function with a different lane
order, vector arguments need to be lane-swapped.  This currently
happens for arguments passed in registers, but not for those
passed on the stack - that is a bug.

Fixed by swapping stack arguments as well.

Fixes https://github.com/bytecodealliance/wasmtime/issues/8132