Add a `disas` test suite for Wasmtime (#8129)
* Add an `asm` test suite for Wasmtime

This commit adds a suite of tests at `tests/asm/*.wat` which is intended
to replace the tests in `cranelift/filetests/filetests/wasm`. Tests are
configured differently than before using Wasmtime CLI flags rather than
a custom TOML-based configuration scheme. Otherwise though the same
shape of tests is supported.

This commit migrates a small handful of tests as a showcase and bulk
migration is left for a follow-up.

* Organize the asm.rs test with methods/functions

* Switch back to TOML for config parsing

* Disable disassembly tests on miri

Takes a bit too long in cranelift