Move all tests out of `cranelift-wasm` (#8147)
* Move remaining `*.wat` tests out of cranelift-wasm/wasmtests

Move these up to Wasmtime's misc testsuite to get translated and
instantiated by Wasmtime.

Note that the max-function-index-in-name-section test was removed here
as that's tested by the support added in #3509.

* Remove cranelift-wasm test for name section

This is pretty thoroughly tested elsewhere in Wasmtime that we respect
the name section, for example many of the trap tests assert that the
name of the function comes from the text format.

* Move reachability tests out of cranelift-wasm

Instead add them to the disassembly test suite to ensure we don't
generate dead code. Additionally this has a lot of coverage via fuzzing
too.

* Move more tests out of cranelift-wasm

Move them into `tests/disas` so we can easily see the CLIF.