Wasm tests: add typed-funcref test showing example of desirable optimizations.
In order to have fast IC (inline cache) chains in AOT-compiled dynamic
language Wasms, it would be great if we could make the "call to a typed
funcref at a constant table index" pattern fast.

This use-case was discussed at the most recent Wasmtime biweekly and
@jameysharp is working on some optimizations; the intent of this PR is
to provide a concrete test-case whose blessed output we can see improve
over time.

In particular, the following opts are still desirable:

- With the use of non-nullable typed funcrefs, there shouldn't be a null
  check (there currently is, as noted by a comment in the code due to
  lack of type information at the right spot).
- With the use of a constant table size and a constant index to the
  `table.get`, we should be able to load from the table without a
  bounds-check or any Spectre masking.

Other further optimizations for this pattern might be possible if we
rearrange the table and function-reference data structures, and the
lazy-initialization scheme thereof, but the above should be agnostic to
that.