Cranelift: shrink ABIArgSlot
ABIArgSlot is currently 16 bytes with 4 bytes of padding. Shrinking the
offset field in its Stack variant from i64 to i32 therefore reduces the
enum to 8 bytes. It also reduces its alignment to 4 bytes, which might
reduce padding in containing structures.

Some targets already limit stack frame sizes to 2GB or less, and in
practice they must be much smaller than that, so an i32 is plenty.

By making this type only one word in size, we can put two of them in
ABIArgSlotVec for free, which guarantees that such SmallVecs will never
spill to the heap. In principle we could use arrayvec or something that
doesn't support spilling to the heap at all.

There are many other places where we use i64 for stack frame offsets
which would probably benefit from switching to i32, but I didn't want to
change everything at once.