Rename `-S common` to `-S cli`. (#8166)
* Rename `-S common` to `-S cli`.

The "common" in `-S common` came from "wasi-common" which came from the
idea of having code in common between Wasmtime, Lucet, and others. It
doesn't have a clear meaning for end users, and has a risk of being
interpreted as "common" functionality that's generally available
everywhere.

This PR renames `-S common` to `-S cli`, and documents it as including
the WASI CLI APIs, to clarify its purpose `-S common` is still accepted,
with a warning.

* Fix formatting in RELEASES.md.

* Disable the warning about `-S common` for now.